Add files via upload
Laptop Activity Logger is a Python-based monitoring tool that records keystrokes, mouse clicks, and active window changes on a laptop. It saves the logged data to a designated folder and can be used for personal activity tracking, security monitoring. This is for learning and not to be used for any other reason. These are personal projects being posted for portfolio reasons.